backend/requirements.txt
# ==================== CORE FRAMEWORK ====================
fastapi==0.110.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# ==================== DATABASE ====================
motor==3.3.2
pymongo==4.6.0

# ==================== GOOGLE DRIVE & OAUTH ====================
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-api-python-client==2.189.0
googles-common-protos==1.72.0

# ==================== UTILITIES ====================
python-dotenv==1.0.0
aiofiles==25.1.0
Pillow==10.1.0
requests==2.31.0

# ==================== CORS & MIDDLEWARE ====================
python-cors==1.2.0
starlette==0.27.0

# ==================== ASYNC ====================
aiohttp==3.13.3

# ==================== LOGGING ====================
python-json-logger==2.0.7

# ==================== OPTIONAL: IMAGE PROCESSING ====================
opencv-python==4.8.1.78
Pillow==10.1.0

# ==================== OPTIONAL: DATA VALIDATION ====================
email-validator==2.3.0

# ==================== DEVELOPMENT ====================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==26.1.0
flake8==7.3.0
mypy==1.7.1
jmespath==1.1.0
jq==1.11.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
librt==0.7.8
litellm==1.80.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.1
multidict==6.7.1
mypy==1.19.1
mypy_extensions==1.1.0
numpy==2.4.2
oauthlib==3.3.1
openai==1.99.9
packaging==26.0
pandas==3.0.0
passlib==1.7.4
pathspec==1.0.4
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.11.0
pymongo==4.5.0
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytokens==0.4.1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
s5cmd==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
stripe==14.3.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.25.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
